Backend/requirements.txt
flask
sqlalchemy
flask_sqlalchemy
flask_cors
flask-Migrate
PyJWT
openai
dotenv
pymysql
requests
email-validator
google-generativeai
cryptography
google-auth
google-auth-oauthlib
python-magic-bin
pydantic

#file_readers
python-docx
python-pptx
openpyxl
pypdf2
odfpy
pyyaml
beautifulsoup4
striprtf
defusedxml

# IA advanced tools
chromadb
pypdf
tiktoken # for token counting
sentence-transformers